pyproject.toml
[project]
name = "basic-crud-operations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
